requirements.txt
black==23.7.0
bokeh==3.3.4
chardet==5.1.0
clip @ git+https://github.com/openai/CLIP.git
einops>=0.6.1
fairscale>=0.4.13
fire>=0.5.0
fsspec>=2023.6.0
invisible-watermark>=0.2.0
joblib==1.3.2
kornia==0.6.9
lovely_numpy==0.2.11
lovely-tensors==0.1.15
matplotlib>=3.7.2
natsort>=8.4.0
ninja>=1.11.1
numpy>=1.24.4
omegaconf>=2.3.0
open3d==0.18.0
open-clip-torch==2.24.0
opencv-python==4.6.0.66
pandas>=2.0.3
panda3d==1.10.14
pillow>=9.5.0
pin==2.7.0
pudb>=2022.1.3
pypng==0.20220715.0
pytorch-lightning==2.0.1
pyyaml>=6.0.1
rembg
roma==1.4.3
scipy>=1.10.1
seaborn==0.13.2
simplejson==3.19.2
streamlit>=0.73.1
tensorboardx==2.6
timm>=0.9.2
tokenizers==0.12.1
torch>=2.0.1
torchaudio>=2.0.2
torchdata==0.6.1
torchmetrics>=1.0.1
torchvision>=0.15.2
tqdm>=4.65.0
transformers==4.19.1
transforms3d==0.4.1
trimesh==4.1.6
triton==2.0.0
urllib3<1.27,>=1.25.4
wandb>=0.15.6
webdataset>=0.2.33
wheel>=0.41.0
xformers>=0.0.20
gradio
streamlit-keyup==0.2.0

